services/simulador/requirements.txt
kafka-python-ng==2.2.3
faker==22.0.0
python-dotenv==1.0.0
pyinstaller==6.3.0